Renamed file from index to PlanningOverview
