+++ src/views/index.tsx
@@ -1,2 +1,2 @@
 export { Editor } from './Editor'
-export { PlanningOverview } from './PlanningOverview'
+export { PlanningOverview } from './PlanningOverview/PlanningOverview'
